+++ pkg/inst-api/utils/scope.go
@@ -39,4 +39,4 @@ const GOMICRO_SERVER_SCOPE_NAME = "pkg/rules/gomicro/server/gomicro_server_setup
 const TRPCGO_CLIENT_SCOPE_NAME = "pkg/rules/trpc/trpc_client_setup.go"
 const TRPCGO_SERVER_SCOPE_NAME = "pkg/rules/trpc/trpc_server_setup.go"
 const LANGCHAIN_SCOPE_NAME = "pkg/rules/langchain/setup.go"
-const AMQP091GO_SCOPE_NAME = "pkg/rules/amqp091go/setup.go"
+const AMQP091_SCOPE_NAME = "pkg/rules/amqp091/setup.go"
