+++ pkg/rules/amqp091/rabbitmq_data_type.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package amqp091go
+package amqp091
 
 type RabbitRequest struct {
 	operationName   string
