+++ test/amqp091.go
@@ -23,7 +23,7 @@ import (
 )
 
 const rabbitmq_dependency_name = "https://github.com/rabbitmq/amqp091-go"
-const rabbitmq_module_name = "amqp091go"
+const rabbitmq_module_name = "amqp091"
 
 func init() {
 	TestCases = append(TestCases,
@@ -35,7 +35,7 @@ func init() {
 func TestRabbitMQCascading(t *testing.T, env ...string) {
 	rabbitC, port := initRabbitMQContainer()
 	defer testcontainers.CleanupContainer(t, rabbitC)
-	UseApp("amqp091go/v1.10.0")
+	UseApp("amqp091/v1.10.0")
 	RunGoBuild(t, "go", "build", "test_mq_cascading.go", "base.go")
 	env = append(env, "RabbitMQ_PORT="+port.Port())
 	RunApp(t, "test_mq_cascading", env...)
@@ -43,7 +43,7 @@ func TestRabbitMQCascading(t *testing.T, env ...string) {
 func TestRabbitMQNOCascading(t *testing.T, env ...string) {
 	rabbitC, port := initRabbitMQContainer()
 	defer testcontainers.CleanupContainer(t, rabbitC)
-	UseApp("amqp091go/v1.10.0")
+	UseApp("amqp091/v1.10.0")
 	RunGoBuild(t, "go", "build", "test_mq_no_cascading.go", "base.go")
 	env = append(env, "RabbitMQ_PORT="+port.Port())
 	RunApp(t, "test_mq_no_cascading", env...)
