+++ test/amqp091/v1.10.0/go.mod
